requirements.txt
asgiref==3.8.1
autopep8==1.6.0
certifi==2024.2.2
charset-normalizer==3.3.2
crispy-bootstrap4==2024.1
distlib==0.3.8
dj-stripe==2.8.4
Django==5.0.6
django-crispy-forms==2.1
django-filter==24.2
django-formtools==2.3
django-ratings==0.3.7
djangorestframework==3.15.1
filelock==3.13.3
idna==3.7
isort==5.10.1
Markdown==3.6
pillow==10.2.0
platformdirs==4.2.0
pycodestyle==2.8.0
requests==2.31.0
sqlparse==0.4.2
stripe==4.2.0
toml==0.10.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.25.1
